pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vn-pqc-analyzer"
version = "0.3.0"
description = "Assess your system's readiness for post-quantum cryptography migration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nguyen Dong", email = "support@vradar.io" },
]
keywords = [
    "post-quantum",
    "cryptography",
    "pqc",
    "security",
    "scanner",
    "nist",
    "kyber",
    "dilithium",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Networking",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "cryptography>=42.0.0",
    "pyOpenSSL>=24.0.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]
web = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "jinja2>=3.1.0",
]
report = [
    "jinja2>=3.1.0",
    "weasyprint>=61.0",
    "plotly>=5.20.0",
]
benchmark = [
    "liboqs-python>=0.10.0",
    "psutil>=5.9.0",
]
flow = [
    "scapy>=2.5.0",
    "PyYAML>=6.0",
]
flow-extras = [
    "pyshark>=0.6",
]
all = [
    "vn-pqc-analyzer[dev,web,report,benchmark,flow]",
]

[project.scripts]
pqc-analyzer = "src.cli:app"

[project.urls]
Homepage = "https://github.com/xuxu298/PQCAnalyzer"
Documentation = "https://github.com/xuxu298/PQCAnalyzer/tree/main/docs"
Repository = "https://github.com/xuxu298/PQCAnalyzer"
Issues = "https://github.com/xuxu298/PQCAnalyzer/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring network or Docker (deselect with '-m \"not integration\"')",
]
